requirements.txt
annoy==1.17.3
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
comm==0.1.3
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.48.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.60.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpickle==3.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.25.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyfiglet==0.8.post1
Pygments==2.15.1
pynytimes==0.10.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-resize-image==1.1.20
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
stack-data==0.6.2
sympy==1.12
termcolor==1.1.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
